+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- CASMCMS-9211: Improve performance of configuration delete operation.
+
 ### Fixed
 - CASMCMS-9210: Correctly check `additional_inventory` layer when seeing if a source is in use